cpu/esp8266: print_meminfo function added
