+++ cpu/esp8266/tools.c
@@ -20,9 +20,23 @@
 
 #include <stdio.h>
 #include <inttypes.h>
+#include <malloc.h>
 
 #include "esp/common_macros.h"
+#include "sdk/ets.h"
 #include "tools.h"
+
+extern void malloc_stats (void);
+extern unsigned int get_free_heap_size (void);
+extern uint8_t _eheap;  /* end of heap (defined in esp8266.riot-os.app.ld) */
+extern uint8_t _sheap;  /* start of heap (defined in esp8266.riot-os.app.ld) */
+
+void print_meminfo (void)
+{
+    struct mallinfo minfo = mallinfo();
+    ets_printf("heap: %lu (free %lu) byte\n", &_eheap - &_sheap, get_free_heap_size());
+    ets_printf("sysmem: %d (used %d, free %d)\n", minfo.arena, minfo.uordblks, minfo.fordblks);
+}
 
 void esp_hexdump (const void* addr, uint32_t num, char width, uint8_t per_line)
 {
